enhancement: add a way to register a u&p provider
