+++ packages/plugins/users-permissions/server/utils/index.d.ts
@@ -3,6 +3,7 @@ import * as user from '../services/user';
 import * as role from '../services/role';
 import * as jwt from '../services/jwt';
 import * as providers from '../services/providers';
+import * as providersRegistry from '../services/providers-registry';
 import * as permission from '../services/permission';
 
 type S = {
@@ -11,7 +12,7 @@ type S = {
   user: typeof user;
   jwt: typeof jwt;
   providers: typeof providers;
-  ['providers-registry']: typeof providers;
+  ['providers-registry']: typeof providersRegistry;
   permission: typeof permission;
 };
